Refactor: view 폴더 위치 변경 및 에러 처리
- main 파일에서 try-catch 에러처리 \n-output 에러 메세지 출력 함수 추가
