+++ src/domain/Car.js
@@ -1,11 +1,12 @@
+import { ERROR_MESSAGES } from '../utils/constants.js';
+
 class Car {
-  static INVALID_CAR_NAME = "잘못된 자동차 이름입니다.";
   #name;
   #location = 0;
 
   constructor(name) {
     if (!this.isValidName(name)) {
-      throw new Error(Car.INVALID_CAR_NAME);
+      throw new Error(ERROR_MESSAGES.INVALID_CAR_NAME);
     }
     this.#name = name;
   }
