+++ src/main.js
@@ -1,19 +1,23 @@
-import Input from "../view/Input.js";
+import Input from "./view/Input.js";
 import Race from "./domain/Race.js";
-import Output from "../view/Output.js";
+import Output from "./view/Output.js";
 import Car from "./domain/Car.js";
 
 async function main() {
   const input = new Input();
   const output = new Output();
 
-  const carNames = await input.askCarNames();
-  const cars = carNames.map((name) => new Car(name));
+  try {
+    const carNames = await input.askCarNames();
+    const cars = carNames.map((name) => new Car(name));
 
-  const race = new Race(cars);
+    const race = new Race(cars);
 
-  race.start();
-  output.printRaceResult(race.result);
+    race.start();
+    output.printRaceResult(race.result);
+  } catch (error) {
+    output.printErrorMessage(error);
+  }
 }
 
 main();
